+++ tests/test_experiment.py
@@ -3,7 +3,7 @@ import os
 import signal
 import tempfile
 import types
-from datetime import datetime
+from datetime import datetime, timezone
 
 import pytest
 import requests_mock
@@ -397,9 +397,9 @@ def test_can_run_experiment_with_activity_in_dry_mode():
 def test_dry_run_should_not_pause_after():
     experiment = experiments.ExperimentWithLongPause.copy()
     experiment["dry"] = Dry.ACTIVITIES
-    start = datetime.utcnow()
+    start = datetime.now(timezone.utc)
     run_experiment(experiment)
-    end = datetime.utcnow()
+    end = datetime.now(timezone.utc)
 
     experiment_run_time = int((end - start).total_seconds())
     pause_after_duration = int(experiment["method"][1]["pauses"]["after"])
@@ -410,9 +410,9 @@ def test_dry_run_should_not_pause_after():
 def test_actionless_run_should_not_pause_after():
     experiment = experiments.ExperimentWithLongPauseAction.copy()
     experiment["dry"] = Dry.ACTIONS
-    start = datetime.utcnow()
+    start = datetime.now(timezone.utc)
     run_experiment(experiment)
-    end = datetime.utcnow()
+    end = datetime.now(timezone.utc)
 
     experiment_run_time = int((end - start).total_seconds())
     pause_after_duration = int(experiment["method"][1]["pauses"]["after"])
@@ -423,9 +423,9 @@ def test_actionless_run_should_not_pause_after():
 def test_probeless_run_should_not_pause_after():
     experiment = experiments.ExperimentWithLongPause.copy()
     experiment["dry"] = Dry.PROBES
-    start = datetime.utcnow()
+    start = datetime.now(timezone.utc)
     run_experiment(experiment)
-    end = datetime.utcnow()
+    end = datetime.now(timezone.utc)
 
     experiment_run_time = int((end - start).total_seconds())
     pause_after_duration = int(experiment["method"][1]["pauses"]["after"])
@@ -436,9 +436,9 @@ def test_probeless_run_should_not_pause_after():
 def test_pauseless_run_should_not_pause_after():
     experiment = experiments.ExperimentWithLongPause.copy()
     experiment["dry"] = Dry.PAUSE
-    start = datetime.utcnow()
+    start = datetime.now(timezone.utc)
     run_experiment(experiment)
-    end = datetime.utcnow()
+    end = datetime.now(timezone.utc)
 
     experiment_run_time = int((end - start).total_seconds())
     pause_after_duration = int(experiment["method"][1]["pauses"]["after"])
@@ -449,9 +449,9 @@ def test_pauseless_run_should_not_pause_after():
 def test_dry_run_should_not_pause_before():
     experiment = experiments.ExperimentWithLongPauseBefore.copy()
     experiment["dry"] = Dry.ACTIVITIES
-    start = datetime.utcnow()
+    start = datetime.now(timezone.utc)
     run_experiment(experiment)
-    end = datetime.utcnow()
+    end = datetime.now(timezone.utc)
 
     experiment_run_time = int((end - start).total_seconds())
     pause_before_duration = int(experiment["method"][1]["pauses"]["before"])
@@ -463,9 +463,9 @@ def test_actionless_run_should_not_pause_before():
     experiment = experiments.ExperimentWithLongPauseAction.copy()
     experiment["dry"] = Dry.ACTIONS
 
-    start = datetime.utcnow()
+    start = datetime.now(timezone.utc)
     run_experiment(experiment)
-    end = datetime.utcnow()
+    end = datetime.now(timezone.utc)
 
     experiment_run_time = int((end - start).total_seconds())
     pause_before_duration = int(experiment["method"][1]["pauses"]["before"])
@@ -477,9 +477,9 @@ def test_probeless_run_should_not_pause_before():
     experiment = experiments.ExperimentWithLongPauseBefore.copy()
     experiment["dry"] = Dry.PROBES
 
-    start = datetime.utcnow()
+    start = datetime.now(timezone.utc)
     run_experiment(experiment)
-    end = datetime.utcnow()
+    end = datetime.now(timezone.utc)
 
     experiment_run_time = int((end - start).total_seconds())
     pause_before_duration = int(experiment["method"][1]["pauses"]["before"])
@@ -491,9 +491,9 @@ def test_pauseless_run_should_not_pause_before():
     experiment = experiments.ExperimentWithLongPauseBefore.copy()
     experiment["dry"] = Dry.PAUSE
 
-    start = datetime.utcnow()
+    start = datetime.now(timezone.utc)
     run_experiment(experiment)
-    end = datetime.utcnow()
+    end = datetime.now(timezone.utc)
 
     experiment_run_time = int((end - start).total_seconds())
     pause_before_duration = int(experiment["method"][1]["pauses"]["before"])